refactor: move transform to `test/util`
