+++ src/test/util.ts
@@ -1,6 +1,9 @@
+import less from 'less';
 import postcss, { Root, Rule, AtRule, Declaration } from 'postcss';
 import { ClassName } from 'postcss-selector-parser';
+import sass from 'sass';
 import { collectNodes } from '../../src/postcss';
+import { Transformer } from '../loader';
 
 export function createRoot(code: string, from?: string): Root {
   return postcss.parse(code, { from: from || '/test/test.css' });
@@ -17,3 +20,17 @@ export function createClassSelectors(root: Root): { rule: Rule; classSelector: C
 export function createComposesDeclarations(root: Root): Declaration[] {
   return collectNodes(root).composesDeclarations;
 }
+
+export const transform: Transformer = async (source: string, from: string) => {
+  if (from.endsWith('.scss')) {
+    const result = sass.compile(from, { sourceMap: true });
+    return { css: result.css, map: result.sourceMap!, dependencies: result.loadedUrls };
+  } else if (from.endsWith('.less')) {
+    const result = await less.render(source, {
+      filename: from,
+      sourceMap: {},
+    });
+    return { css: result.css, map: result.map, dependencies: result.imports };
+  }
+  return false;
+};
